Refactor classification e2e tests
