+++ cypress/integration/classification.common.ts
@@ -1,122 +1,18 @@
-import { gql } from "@apollo/client";
-
-import { distantDatabaseClient as client } from "../../typescript/web/src/connectors/apollo-client/client";
-
-const createDataset = async (name: string) => {
-  const mutationResult = await client.mutate({
-    mutation: gql`
-      mutation createDataset($name: String) {
-        createDataset(
-          data: { name: $name, workspaceSlug: "cypress-test-workspace" }
-        ) {
-          id
-          slug
-        }
-      }
-    `,
-    variables: {
-      name,
-    },
-  });
-
-  const {
-    data: {
-      createDataset: { id, slug },
-    },
-  } = mutationResult;
-
-  return { id, slug };
-};
-
-async function createImage(url: string, datasetId: string) {
-  const mutationResult = await client.mutate({
-    mutation: gql`
-      mutation createImage($url: String, $datasetId: ID!) {
-        createImage(data: { url: $url, datasetId: $datasetId }) {
-          id
-          name
-          width
-          height
-          url
-        }
-      }
-    `,
-    variables: {
-      datasetId,
-      url,
-    },
-  });
-
-  const {
-    data: { createImage: image },
-  } = mutationResult;
-
-  return image;
-}
-
-const createLabelClass = async (
-  name: String,
-  color = "#ffffff",
-  datasetId: string
-) => {
-  const {
-    data: {
-      createLabelClass: { id },
-    },
-  } = await client.mutate({
-    mutation: gql`
-      mutation createLabelClass(
-        $name: String!
-        $color: String!
-        $datasetId: ID!
-      ) {
-        createLabelClass(
-          data: { name: $name, color: $color, datasetId: $datasetId }
-        ) {
-          id
-          name
-          color
-        }
-      }
-    `,
-    variables: {
-      name,
-      color,
-      datasetId,
-    },
-  });
-
-  return id;
-};
-
-describe("Classification", () => {
-  let datasetId: string;
-  let datasetSlug: string;
-  let imageId: string;
-  beforeEach(() => {
-    cy.setCookie("consentedCookies", "true");
-    cy.task("performLogin").then((token) => {
-      cy.setCookie("next-auth.session-token", token as string);
-    });
-    cy.task("createWorkspaceAndDatasets").then(async () => {
-      const createResult = await createDataset("cypress test dataset");
-      datasetId = createResult.id;
-      datasetSlug = createResult.slug;
-
-      const { id } = await createImage(
-        "https://images.unsplash.com/photo-1579513141590-c597876aefbc?auto=format&fit=crop&w=882&q=80",
-        datasetId
-      );
-      imageId = id;
-
-      await createLabelClass("Rocket", "#F87171", datasetId);
-    });
-  });
-
+export const declareClassificationTests = ({
+  workspaceSlug,
+  getDatasetSlug,
+  getImageId,
+}: {
+  workspaceSlug: string;
+  getDatasetSlug: () => string;
+  getImageId: () => string;
+}) => {
   it("switches between drawing tools", () => {
+    const datasetSlug = getDatasetSlug();
+    const imageId = getImageId();
     // See https://docs.cypress.io/guides/core-concepts/conditional-testing#Welcome-wizard
     cy.visit(
-      `/cypress-test-workspace/datasets/${datasetSlug}/images/${imageId}?modal-welcome=closed&modal-update-service-worker=update`
+      `/${workspaceSlug}/datasets/${datasetSlug}/images/${imageId}?modal-welcome=closed&modal-update-service-worker=update`
     );
     cy.get('[aria-label="loading indicator"]').should("not.exist");
     cy.get('[aria-label="Drawing classification tool"]').should("not.exist");
@@ -136,9 +32,11 @@ describe("Classification", () => {
   });
 
   it("add classses and remove them", () => {
+    const datasetSlug = getDatasetSlug();
+    const imageId = getImageId();
     // See https://docs.cypress.io/guides/core-concepts/conditional-testing#Welcome-wizard
     cy.visit(
-      `/cypress-test-workspace/datasets/${datasetSlug}/images/${imageId}?modal-welcome=closed&modal-update-service-worker=update`
+      `/${workspaceSlug}/datasets/${datasetSlug}/images/${imageId}?modal-welcome=closed&modal-update-service-worker=update`
     );
 
     // Switch to classification tool
@@ -286,4 +184,4 @@ describe("Classification", () => {
     cy.focused().type("1");
     cy.get('[aria-label="Classification tag: Rocket"]').should("not.exist");
   });
-});
+};
